Refectoring the ServerProcess with Registered processes
