+++ lib/todo/server_process.ex
@@ -1,6 +1,7 @@
 defmodule ServerProcess do
   def start(callback_module) do
     spawn(fn ->
+      Process.register(self(), :todo_server)
       initial_state = callback_module.init()
       loop(callback_module, initial_state)
     end)
