+++ lib/todo/simple_todo.ex
@@ -5,12 +5,12 @@ defmodule TodoServer do
 
   # Instruction on how to use this function to add an entry
   # TodoServer.add_entry(%{date: ~D[2023-12-19], title: "Dentist"})
-  def add_entry(todo_server, new_entry) do
-    send(todo_server, {:add_entry, new_entry})
+  def add_entry(new_entry) do
+    send(:todo_server, {:add_entry, new_entry})
   end
 
-  def entries(todo_server, date) do
-    ServerProcess.call(todo_server, {:entries, date})
+  def entries(date) do
+    ServerProcess.call(:todo_server, {:entries, date})
   end
 
   def init do
@@ -23,66 +23,6 @@ defmodule TodoServer do
 
   def handle_call({:entries, date}, todo_list) do
     {TodoList.entries(todo_list, date), todo_list}
-  end
-end
-
-defmodule TodoList do
-  # Struct that describes the To-do list
-  defstruct next_id: 1, entries: %{}
-
-  # Interativelly create a new TodoList itens from a raw list of Entry instances
-  # into an instance of the struct %TodoList struct relying on the Enum.reduce/3
-  def new(entries \\ []) do
-    Enum.reduce(
-      entries,
-      %TodoList{},
-      fn entry, todo_list_acc ->
-        add_entry(todo_list_acc, entry)
-      end
-    )
-  end
-
-  # The two arguments are:
-  # a to-do list instance, generated by the previous new()
-  # a Map describing the entry
-  #
-  # Example:
-  # simple_map = %{date: ~D[2023-12-19], title: "Play"}
-  # SimpleTodo.add_entry(%{}, simple_map)
-
-  def add_entry(todo_list, entry) do
-    entry = Map.put(entry, :id, todo_list.next_id)
-
-    new_entries =
-      Map.put(
-        todo_list.entries,
-        todo_list.next_id,
-        entry
-      )
-
-    %TodoList{todo_list | entries: new_entries, next_id: todo_list.next_id + 1}
-  end
-
-  def entries(todo_list, date) do
-    todo_list.entries
-    |> Map.values()
-    |> Enum.filter(fn entry -> entry.date == date end)
-  end
-
-  def update_entry(todo_list, entry_id, updater_fun) do
-    case Map.fetch(todo_list.entries, entry_id) do
-      :error ->
-        todo_list
-
-      {:ok, old_entry} ->
-        new_entry = updater_fun.(old_entry)
-        new_entries = Map.put(todo_list.entries, new_entry.id, new_entry)
-        %TodoList{todo_list | entries: new_entries}
-    end
-  end
-
-  def delete_entry(todo_list, entry_id) do
-    %TodoList{todo_list | entries: Map.delete(todo_list.entries, entry_id)}
   end
 end
